Aggiornate tipologie di modello User

+++ app/src/main/java/com/rapnap/panpar/repository/AuthRepository.kt
@@ -8,12 +8,12 @@ import com.google.firebase.Timestamp
 import com.google.firebase.auth.AuthCredential
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
+import com.google.firebase.firestore.GeoPoint
+import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.rapnap.panpar.model.Tipologia
 import com.rapnap.panpar.model.Utente
 import java.util.*
-import com.google.firebase.firestore.GeoPoint
-import com.google.firebase.firestore.ktx.firestore
 
 
 class AuthRepository {
@@ -129,7 +129,7 @@ class AuthRepository {
                         val location = document.data?.get("location") as GeoPoint
                         user.location = location.toString()
                         user.punteggio = document.data?.get("punteggio") as Long
-                        user.rating = document.data?.get("rating") as Long
+                        user.rating = document.data?.get("rating") as Double
                         user.isNew = false
                         Log.d(TAG, "User data created: ${user.toString()}")
